pyproject.toml
[project]
name = "fightgamecharbot"
version = "0.1.0"
description = "Chat bot for fight game using llm"
requires-python = ">=3.12"
dependencies = [
    "grpcio==1.73.0rc1",
    "grpcio-tools==1.73.0rc1",
    "langchain-google-genai==2.1.5",
    "langgraph==0.4.8",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
